User commit

+++ src/main/java/com/example/demo/entity/ValidationEntity.java
@@ -28,7 +28,7 @@ public int getId(){
     return id;
 }
 
-public void setId(Long id){
+public void setId(int id){
     this.id=id;
 }
 public String getUsername(){
